New approach to units

+++ SportClubsChallenges.Model/Dto/PeriodStatsDto.cs
@@ -4,10 +4,10 @@
     {
         public long Count { get; set; }
 
-        public long Distance { get; set; }
+        public int Distance { get; set; }
 
-        public long Duration { get; set; }
+        public int Duration { get; set; }
 
-        public long Elevation { get; set; }
+        public int Elevation { get; set; }
     }
 }
